Input v1, Time v1, Updater and OS detection
The Input system is working on windows, however it is really easy to adapt the structure of the wrappers (create a IInputWrapper for each OS with its own KeyPress detection method).

The time is able to count since the run time of the engine, the updater is computing the DeltaTime (time the frame takes to perform).

The FixedTime is not implemented yet.