Vectors
Vectors are now a thing, with 2D, 3D, 4D and ND.

2D, 3D and 4D ones have all of their components accessible and addable, multiply-able, subtract-able..
Their components are in available in whatever order like ".xy" or ".zwy". However the ND vectors do not have this functionality and their components are accessible through an indexer: vectorNd[dimension] = value.

It is also possible to create the vectors in the usual way (x,y,z) but also by combining vectors with values, like Vector2D + N. Vectors can also be created with only one value that will fill all the components.